requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pgvector
psycopg2-binary
langchain
langchain-openai
langchain-community
openai
python-dotenv
pydantic